+++ packages/acceptance-tests/pkg-tests-specs/sources/node-modules.test.ts
@@ -1123,4 +1123,49 @@ describe(`Node_Modules`, () => {
       }
     )
   );
+
+  test(`should honor transparently nmMode: hardlinks during subsequent installs`,
+    makeTemporaryEnv(
+      {
+        workspaces: [`ws1`, `ws2`, `ws3`],
+        dependencies: {
+          [`no-deps`]: `1.0.0`,
+        },
+      },
+      async ({path, run}) => {
+        await writeJson(`${path}/ws1/package.json`, {
+          dependencies: {
+            [`no-deps`]: `1.0.0`,
+          },
+        });
+
+        await writeJson(`${path}/ws2/package.json`, {
+          dependencies: {
+            [`no-deps`]: `2.0.0`,
+          },
+        });
+
+        await writeJson(`${path}/ws3/package.json`, {
+          dependencies: {
+            [`no-deps`]: `2.0.0`,
+          },
+        });
+
+        await writeFile(`${path}/.yarnrc.yml`, `nodeLinker: node-modules\nnmMode: hardlinks\n`);
+        await run(`install`);
+
+        expect(await xfs.statPromise(`${path}/ws3/node_modules/no-deps/package.json` as PortablePath)).toMatchObject({nlink: 2});
+
+        await writeFile(`${path}/.yarnrc.yml`, `nodeLinker: node-modules\nnmMode: classic\n`);
+        await run(`install`);
+
+        expect(await xfs.statPromise(`${path}/ws3/node_modules/no-deps/package.json` as PortablePath)).toMatchObject({nlink: 1});
+
+        await writeFile(`${path}/.yarnrc.yml`, `nodeLinker: node-modules\nnmMode: hardlinks\n`);
+        await run(`install`);
+
+        expect(await xfs.statPromise(`${path}/ws3/node_modules/no-deps/package.json` as PortablePath)).toMatchObject({nlink: 2});
+      }
+    )
+  );
 });